# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Einstein API",
        "description": "API para gerenciamento de clientes e endereços",
        "contact": {
            "email": "contato@einstein.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor de desenvolvimento"
        }
    ],
    "paths": {
...
        "/customers": {
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "Listar todos os clientes",
                "operationId": "30ce4124bc07a0d013b69974ff225a5c",
                "responses": {
                    "200": {
                        "description": "Lista de clientes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Customer"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Customers"
                ],
                "summary": "Criar um novo cliente",
                "operationId": "8350a1a36d2b5522ea4158253eff28a1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Customer"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cliente criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 201
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Cliente criado com sucesso"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Customer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/customers/{id}": {
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "Exibir um cliente específico",
                "operationId": "bb85afd644c1b965544c2f0e51f408b9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cliente encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Customer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cliente não encontrado"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Customers"
                ],
                "summary": "Atualizar um cliente específico",
                "operationId": "e33f1a95a01adac2456879f7df17acdb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Customer"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cliente atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Cliente atualizado com sucesso"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Customer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cliente não encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Customer": {
                "title": "Customer",
                "required": [
                    "id",
                    "name",
                    "email",
                    "cpf"
                ],
                "properties": {
                    "id": {
                        "description": "ID do cliente",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Nome do cliente",
                        "type": "string",
                        "example": "Maria Oliveira"
                    },
                    "email": {
                        "description": "Email do cliente",
                        "type": "string",
                        "format": "email",
                        "example": "maria@email.com"
                    },
                    "cpf": {
                        "description": "CPF do cliente",
                        "type": "string",
                        "example": "123.456.789-00"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}